Added victory music, changed level state managers
