+++ .gitignore
@@ -23,3 +23,8 @@ ExportedObj/
 
 # Unity3D Generated File On Crash Reports
 sysinfo.txt
+
+# Built project files
+/PoolGame_Data/
+*.pdb
+*.exe
