--- Assets/Standard Assets/Scripts/General Scripts/ActivateTrigger.cs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c8634e062924929664361c08745211fb
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
