--- Assets/Standard Assets/Scripts/General Scripts/ButtonToNext.cs
@@ -1,22 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class ButtonToNext : MonoBehaviour
-{
-
-    public class ButtonNextLevel : MonoBehaviour
-    {
-
-        public void NextLevelButton(int index)
-        {
-
-            Application.LoadLevel(index);
-        }
-
-        public void NextLevelButton(string levelName)
-        {
-
-            Application.LoadLevel(levelName);
-        }
-    }
-}
