--- Assets/Standard Assets/Scripts/General Scripts/ButtonToNext.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d8b50158c7df26e4db211bf991222707
-timeCreated: 1449608800
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
