--- Assets/Standard Assets/Scripts/General Scripts/DragRigidbody.js.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cae09c86ad1122e5d0006ac98d76c639
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
