+++ Assets/Standard Assets/Scripts/General Scripts/NormalPoolBall.cs
@@ -6,7 +6,6 @@ public class NormalPoolBall : AbstractPoolBall
     {
         // on score normal pool balls add 1 to the total score and
         // are removed from play
-        Debug.Log(gameObject);
         gameObject.SetActive(false);
 
         scoreboard.UpdateScore(1);
