+++ Assets/Standard Assets/Scripts/General Scripts/myRigidBody.physicsMaterial2D
@@ -6,5 +6,5 @@ PhysicsMaterial2D:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_Name: myRigidBody
-  friction: .400000006
-  bounciness: 2
+  friction: 0.4
+  bounciness: 0.9
